requirements.txt
tweepy
pyyaml
boto3
elasticsearch
requests_aws_sign